A non-empty zero-indexed matrix A consisting of N rows and M columns of integers is given. A pair (P, Q) is called an equilibrium point of matrix A if the following conditions are met:
	0<= P < N;
	0<= Q < M;
	the sum of elements in rows above the P-th row is equal to the sum of
elements in rows below the P-th row;
		the sum of elements in columns to the left of the Q-th column is
equal to the sum of elements in columns to the right of the Q-th column.

For example, consider matrix A consisting of five rows and three
columns, such that:

A[0][0] = 2	A[0][1] = 7	A[0][2] = 5
A[1][0] = 3 	A[1][1] = 1 	A[1][2] = 1
A[2][0] = 2		A[2][1] = 1 	A[2][2] = -7
A[3][0] = 0 	A[3][1] = 2		A[3][2] = 1
A[4][0] = 1 	A[4][1] = 6 	A[4][2] = 8

Pair(1, 1) is an equilibrium point of A because:
	the sum of elements in 0th row  is equal to the sum of elements in the
2nd, 3rd and 4th rows(sum is 14).
	the sum of elements in 0th column  is equal to the sum of elements in
the 2nd column(sum is 8).

Write a function to find the numbers of total equilibrium point.